app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Francois van Kempen & Sam Kanu", email = "francois@francoisvankempen.com"},
    {name = "samarpro", email = "samarpangupta123@gmail.com"},
]
dependencies = [
    "composio>=0.11.3",
    "pydantic-ai>=1.68.0",
    "fastmcp>=3.1.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.35.0",
    "asyncpg>=0.31.0",
    "sqlalchemy>=2.0.48",
    "greenlet>=3.2.4",
    "pydantic-settings>=2.13.1",
    "logfire>=4.0.0",
    "stamina>=25.2.0",
    "redis[hiredis]>=7.3.0",
    "python-telegram-bot>=22.6",
    "twilio>=9.10.3",
    "docker>=7.1.0",
    "alembic>=1.18.4",
]
requires-python = "==3.14.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[tool.pdm.scripts]
chat = "python -m nightowl.cli"
migrate = "python -m alembic upgrade head"
migration = "python -m alembic revision --autogenerate -m"
rollback = "python -m alembic downgrade -1"

[dependency-groups]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["nightowl"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = ["ignore::logfire._internal.config.LogfireNotConfiguredWarning"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
